Renamed data type files.
