+++ apps/frontend/src/types/bundle.ts
