+++ apps/frontend/src/types/investigation.ts
